+++ inc/class.rateit.php
@@ -2,7 +2,7 @@
 # -- BEGIN LICENSE BLOCK ----------------------------------
 # This file is part of rateIt, a plugin for Dotclear 2.
 #
-# Copyright(c) 2014-2015 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
+# Copyright(c) 2014-2016 Nicolas Roudaire <nikrou77@gmail.com> http://www.nikrou.net
 #
 # Copyright (c) 2009-2010 JC Denis and contributors
 # jcdenis@gdwd.com
@@ -221,7 +221,7 @@ class rateIt
 	{
 		$params['columns'][] = 'COUNT(rateit_id) as rateit_count';
 
-		if ($this->con->driver() == 'mysql')
+		if ($this->con->driver() == 'mysqli')
 		{
 			$params['from'] = 'INNER JOIN '.$this->table.' ON CAST(P.post_id as char)=rateit_id ';
 		}